修复数据库脚本

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,6 +1,6 @@
 create table user
 (
-    id int auto_increment primary key not null,
+    id bigint auto_increment primary key not null,
     account_id varchar(100),
     name varchar(50),
     token varchar(36),
